update: 성별 열거형 타입을 Member 내부 객체에서 외부로 이동 및 Converter
