+++ src/main/java/com/LetMeDoWith/LetMeDoWith/enums/converter/member/GenderConverter.java
@@ -1,0 +1,11 @@
+package com.LetMeDoWith.LetMeDoWith.enums.converter.member;
+
+import com.LetMeDoWith.LetMeDoWith.enums.converter.AbstractConverter;
+import com.LetMeDoWith.LetMeDoWith.enums.member.Gender;
+
+public class GenderConverter extends AbstractConverter<Gender> {
+    
+    public GenderConverter() {
+        super(Gender.class);
+    }
+}
